WIP: split Address / BootProto
